--- a/benchmark/bufferWithCount/bufferWithCount.php
+++ b/benchmark/bufferWithCount/bufferWithCount.php
@@ -1,0 +1,9 @@
+<?php
+
+use Rx\Observable;
+
+return function() use ($dummyObserver) {
+    Observable::range(0, 25)
+        ->bufferWithCount(5)
+        ->subscribe($dummyObserver);
+};