--- a/benchmark/bufferWithCount/bufferWithCount_eventloop.php
+++ b/benchmark/bufferWithCount/bufferWithCount_eventloop.php
@@ -1,0 +1,16 @@
+<?php
+
+use Rx\Observable;
+use Rx\Scheduler\EventLoopScheduler;
+use React\EventLoop\StreamSelectLoop;
+
+$loop = new StreamSelectLoop();
+$scheduler = new EventLoopScheduler($loop);
+
+return function() use ($dummyObserver, $scheduler, $loop) {
+    Observable::range(0, 25)
+        ->bufferWithCount(5)
+        ->subscribe($dummyObserver, $scheduler);
+
+    $loop->run();
+};