--- a/benchmark/catch/catch.php
+++ b/benchmark/catch/catch.php
@@ -1,0 +1,12 @@
+<?php
+
+use Rx\Observable;
+
+$source = Observable::error(new \Exception('error'))
+    ->catchError(function() {
+        return Observable::just(25);
+    });
+
+return function() use ($source) {
+    return $source;
+};