--- a/benchmark/combineLatest/combineLatest.php
+++ b/benchmark/combineLatest/combineLatest.php
@@ -3,7 +3,9 @@
 use Rx\Observable;
 
 $source = Observable::range(0, 25)
-    ->concat(Observable::range(0, 25));
+    ->combineLatest([Observable::range(0, 25)], function($a, $b) {
+        return $a + $b;
+    });
 
 return function() use ($source) {
     return $source;