--- a/benchmark/combineLatest/combineLatest_eventloop.php
+++ b/benchmark/combineLatest/combineLatest_eventloop.php
@@ -7,10 +7,12 @@ use React\EventLoop\StreamSelectLoop;
 $loop = new StreamSelectLoop();
 $scheduler = new EventLoopScheduler($loop);
 
-$source = Observable::range(0, 500, $scheduler)
-    ->takeLast(50);
+$source = Observable::range(0, 25, $scheduler)
+    ->combineLatest([Observable::range(0, 25, $scheduler)], function($a, $b) {
+        return $a + $b;
+    });
 
-$factory = function() use ($source) {
+$factory =  function() use ($source) {
     return $source;
 };