--- a/benchmark/concat/concat.php
+++ b/benchmark/concat/concat.php
@@ -1,0 +1,10 @@
+<?php
+
+use Rx\Observable;
+
+$source = Observable::range(1, 25)
+    ->concat(Observable::range(1, 25));
+
+return function() use ($source) {
+    return $source;
+};