--- a/benchmark/concat/concat_eventloop.php
+++ b/benchmark/concat/concat_eventloop.php
@@ -7,10 +7,10 @@ use React\EventLoop\StreamSelectLoop;
 $loop = new StreamSelectLoop();
 $scheduler = new EventLoopScheduler($loop);
 
-$source = Observable::just(25)
-    ->delay(0, $scheduler);
+$source = Observable::range(1, 25, $scheduler)
+    ->concat(Observable::range(1, 25, $scheduler));
 
-$factory = function() use ($source, $scheduler) {
+$factory = function() use ($source) {
     return $source;
 };