--- a/benchmark/concatAll/concatAll_eventloop.php
+++ b/benchmark/concatAll/concatAll_eventloop.php
@@ -7,10 +7,13 @@ use React\EventLoop\StreamSelectLoop;
 $loop = new StreamSelectLoop();
 $scheduler = new EventLoopScheduler($loop);
 
-$source = Observable::just(25)
-    ->delay(0, $scheduler);
+$source = Observable::range(0, 25, $scheduler)
+    ->map(function() use ($scheduler) {
+        return Observable::range(0, 25, $scheduler);
+    })
+    ->concatAll();
 
-$factory = function() use ($source, $scheduler) {
+$factory = function() use ($source) {
     return $source;
 };