--- a/benchmark/delay/delay.php
+++ b/benchmark/delay/delay.php
@@ -1,0 +1,11 @@
+<?php
+
+use Rx\Observable;
+use Rx\Scheduler\ImmediateScheduler;
+
+$source = Observable::just(25)
+    ->delay(0, new ImmediateScheduler());
+
+return function() use ($source) {
+    return $source;
+};