--- a/benchmark/distinct/distinct.php
+++ b/benchmark/distinct/distinct.php
@@ -1,0 +1,9 @@
+<?php
+
+use Rx\Observable;
+
+return function() use ($dummyObserver) {
+    Observable::range(1, pow(10, 3))
+        ->distinct()
+        ->subscribe($dummyObserver);
+};