--- a/benchmark/distinct/distinct.php
+++ b/benchmark/distinct/distinct.php
@@ -2,12 +2,13 @@
 
 use Rx\Observable;
 
-$source = array_map(function($val) {
+$range = array_map(function($val) {
     return $val % 3;
 }, range(0, 25));
 
-return function() use ($source, $dummyObserver) {
-    Observable::fromArray($source)
-        ->distinct()
-        ->subscribe($dummyObserver);
+$source = Observable::fromArray($range)
+    ->distinct();
+
+return function() use ($source) {
+    return $source;
 };