--- a/benchmark/distinct/distinct_eventloop.php
+++ b/benchmark/distinct/distinct_eventloop.php
@@ -1,20 +1,22 @@
 <?php
 
 use Rx\Observable;
+use Rx\Observable\ArrayObservable;
 use Rx\Scheduler\EventLoopScheduler;
 use React\EventLoop\StreamSelectLoop;
 
 $loop = new StreamSelectLoop();
 $scheduler = new EventLoopScheduler($loop);
 
-$source = array_map(function($val) {
+$range = array_map(function($val) {
     return $val % 3;
 }, range(0, 25));
 
-return function() use ($source, $dummyObserver, $scheduler, $loop) {
-    Observable::fromArray($source)
-        ->distinct()
-        ->subscribe($dummyObserver, $scheduler);
+$source = (new ArrayObservable($range, $scheduler))
+    ->distinct();
 
-    $loop->run();
+$factory = function() use ($source) {
+    return $source;
 };
+
+return [$factory, $loop];