--- a/benchmark/filter/filter.php
+++ b/benchmark/filter/filter.php
@@ -3,9 +3,12 @@
 use Rx\Observable;
 
 return function() use ($dummyObserver) {
-    Observable::range(1, pow(10, 3))
+    Observable::range(0, 50)
         ->filter(function($value) {
             return $value % 2 == 0;
+        })
+        ->filter(function($value) {
+            return $value % 10 == 0;
         })
         ->subscribe($dummyObserver);
 };