--- a/benchmark/filter/filter.php
+++ b/benchmark/filter/filter.php
@@ -2,13 +2,14 @@
 
 use Rx\Observable;
 
-return function() use ($dummyObserver) {
-    Observable::range(0, 50)
-        ->filter(function($value) {
-            return $value % 2 == 0;
-        })
-        ->filter(function($value) {
-            return $value % 10 == 0;
-        })
-        ->subscribe($dummyObserver);
+$source = Observable::range(0, 50)
+    ->filter(function($value) {
+        return $value % 2 == 0;
+    })
+    ->filter(function($value) {
+        return $value % 10 == 0;
+    });
+
+return function() use ($source) {
+    return $source;
 };