--- a/benchmark/filter/filter_eventloop.php
+++ b/benchmark/filter/filter_eventloop.php
@@ -7,15 +7,16 @@ use React\EventLoop\StreamSelectLoop;
 $loop = new StreamSelectLoop();
 $scheduler = new EventLoopScheduler($loop);
 
-return function() use ($dummyObserver, $scheduler, $loop) {
-    Observable::range(0, 50)
-        ->filter(function($value) {
-            return $value % 2 == 0;
-        })
-        ->filter(function($value) {
-            return $value % 10 == 0;
-        })
-        ->subscribe($dummyObserver, $scheduler);
+$source = Observable::range(0, 50, $scheduler)
+    ->filter(function($value) {
+        return $value % 2 == 0;
+    })
+    ->filter(function($value) {
+        return $value % 10 == 0;
+    });
 
-    $loop->run();
+$factory = function() use ($source) {
+    return $source;
 };
+
+return [$factory, $loop];