--- a/benchmark/flatMap/flatMap.php
+++ b/benchmark/flatMap/flatMap.php
@@ -1,0 +1,12 @@
+<?php
+
+use Rx\Observable;
+
+$source = Observable::range(0, 25)
+    ->flatMap(function($x) {
+        return Observable::range($x, 25);
+    });
+
+return function() use ($source) {
+    return $source;
+};