--- a/benchmark/groupBy/groupBy.php
+++ b/benchmark/groupBy/groupBy.php
@@ -4,9 +4,11 @@ use Rx\Observable;
 
 $source = Observable::range(0, 25)
     ->map(function($i) {
-        return $i % 3;
+        return ['key' => $i % 5];
     })
-    ->distinct();
+    ->groupBy(function($item) {
+        return $item['key'];
+    });
 
 return function() use ($source) {
     return $source;