--- a/benchmark/groupBy/groupBy_eventloop.php
+++ b/benchmark/groupBy/groupBy_eventloop.php
@@ -7,8 +7,13 @@ use React\EventLoop\StreamSelectLoop;
 $loop = new StreamSelectLoop();
 $scheduler = new EventLoopScheduler($loop);
 
-$source = Observable::range(0, 500, $scheduler)
-    ->takeLast(50);
+$source = Observable::range(0, 25, $scheduler)
+    ->map(function($i) {
+        return ['key' => $i % 5];
+    })
+    ->groupBy(function($item) {
+        return $item['key'];
+    });
 
 $factory = function() use ($source) {
     return $source;