--- a/benchmark/merge/merge.php
+++ b/benchmark/merge/merge.php
@@ -1,0 +1,10 @@
+<?php
+
+use Rx\Observable;
+
+$source = Observable::range(1, 25)
+    ->merge(Observable::range(1, 25));
+
+return function() use ($source) {
+    return $source;
+};