--- a/benchmark/merge/merge.php
+++ b/benchmark/merge/merge.php
@@ -2,8 +2,8 @@
 
 use Rx\Observable;
 
-$source = Observable::range(1, 25)
-    ->merge(Observable::range(1, 25));
+$source = Observable::range(0, 25)
+    ->merge(Observable::range(0, 25));
 
 return function() use ($source) {
     return $source;