--- a/benchmark/merge/merge_eventloop.php
+++ b/benchmark/merge/merge_eventloop.php
@@ -7,10 +7,10 @@ use React\EventLoop\StreamSelectLoop;
 $loop = new StreamSelectLoop();
 $scheduler = new EventLoopScheduler($loop);
 
-$source = Observable::just(25)
-    ->delay(0, $scheduler);
+$source = Observable::range(0, 250, $scheduler)
+    ->merge(Observable::range(0, 250, $scheduler));
 
-$factory = function() use ($source, $scheduler) {
+$factory = function() use ($source) {
     return $source;
 };