--- a/benchmark/merge/merge_eventloop.php
+++ b/benchmark/merge/merge_eventloop.php
@@ -7,8 +7,8 @@ use React\EventLoop\StreamSelectLoop;
 $loop = new StreamSelectLoop();
 $scheduler = new EventLoopScheduler($loop);
 
-$source = Observable::range(0, 250, $scheduler)
-    ->merge(Observable::range(0, 250, $scheduler));
+$source = Observable::range(0, 25, $scheduler)
+    ->merge(Observable::range(0, 25, $scheduler));
 
 $factory = function() use ($source) {
     return $source;