--- a/benchmark/mergeAll/mergeAll.php
+++ b/benchmark/mergeAll/mergeAll.php
@@ -1,0 +1,13 @@
+<?php
+
+use Rx\Observable;
+
+$source = Observable::range(0, 25)
+    ->map(function() {
+        return Observable::range(0, 25);
+    })
+    ->mergeAll();
+
+return function() use ($source) {
+    return $source;
+};