--- a/benchmark/mergeAll/mergeAll_eventloop.php
+++ b/benchmark/mergeAll/mergeAll_eventloop.php
@@ -7,10 +7,13 @@ use React\EventLoop\StreamSelectLoop;
 $loop = new StreamSelectLoop();
 $scheduler = new EventLoopScheduler($loop);
 
-$source = Observable::just(25)
-    ->delay(0, $scheduler);
+$source = Observable::range(0, 25)
+    ->map(function() {
+        return Observable::range(0, 25);
+    })
+    ->mergeAll();
 
-$factory = function() use ($source, $scheduler) {
+$factory = function() use ($source) {
     return $source;
 };