--- a/benchmark/pluck/pluck.php
+++ b/benchmark/pluck/pluck.php
@@ -2,8 +2,8 @@
 
 use Rx\Observable;
 
-$source = Observable::range(0, 25)
-    ->concat(Observable::range(0, 25));
+$source = Observable::just([1, 2, 3, 4, 5])
+    ->pluck(2);
 
 return function() use ($source) {
     return $source;