--- a/benchmark/reduce/reduce.php
+++ b/benchmark/reduce/reduce.php
@@ -3,7 +3,9 @@
 use Rx\Observable;
 
 $source = Observable::range(0, 25)
-    ->concat(Observable::range(0, 25));
+    ->reduce(function($a, $b) {
+        return $a + $b;
+    });
 
 return function() use ($source) {
     return $source;