--- a/benchmark/reduce/reduce_eventloop.php
+++ b/benchmark/reduce/reduce_eventloop.php
@@ -7,8 +7,10 @@ use React\EventLoop\StreamSelectLoop;
 $loop = new StreamSelectLoop();
 $scheduler = new EventLoopScheduler($loop);
 
-$source = Observable::range(0, 500, $scheduler)
-    ->takeLast(50);
+$source = Observable::range(0, 25, $scheduler)
+    ->reduce(function($a, $b) {
+        return $a + $b;
+    });
 
 $factory = function() use ($source) {
     return $source;