--- a/benchmark/repeat/repeat.php
+++ b/benchmark/repeat/repeat.php
@@ -3,7 +3,7 @@
 use Rx\Observable;
 
 $source = Observable::range(0, 25)
-    ->concat(Observable::range(0, 25));
+    ->repeat(5);
 
 return function() use ($source) {
     return $source;