--- a/benchmark/repeat/repeat_eventloop.php
+++ b/benchmark/repeat/repeat_eventloop.php
@@ -7,8 +7,8 @@ use React\EventLoop\StreamSelectLoop;
 $loop = new StreamSelectLoop();
 $scheduler = new EventLoopScheduler($loop);
 
-$source = Observable::range(0, 500, $scheduler)
-    ->takeLast(50);
+$source = Observable::range(0, 25, $scheduler)
+    ->repeat(5);
 
 $factory = function() use ($source) {
     return $source;