--- a/benchmark/scan/scan.php
+++ b/benchmark/scan/scan.php
@@ -3,7 +3,9 @@
 use Rx\Observable;
 
 $source = Observable::range(0, 25)
-    ->concat(Observable::range(0, 25));
+    ->scan(function($acc, $x) {
+        return $x + $x;
+    });
 
 return function() use ($source) {
     return $source;