--- a/benchmark/scan/scan_eventloop.php
+++ b/benchmark/scan/scan_eventloop.php
@@ -7,8 +7,10 @@ use React\EventLoop\StreamSelectLoop;
 $loop = new StreamSelectLoop();
 $scheduler = new EventLoopScheduler($loop);
 
-$source = Observable::range(0, 500, $scheduler)
-    ->takeLast(50);
+$source = Observable::range(0, 25, $scheduler)
+    ->scan(function($acc, $x) {
+        return $x + $x;
+    });
 
 $factory = function() use ($source) {
     return $source;