--- a/lib/Rx/Operator/DistinctOperator.php
+++ b/lib/Rx/Operator/DistinctOperator.php
@@ -18,13 +18,8 @@ class DistinctOperator implements OperatorInterface
 
     public function __construct(callable $keySelector = null, callable $comparer = null)
     {
-
-        $this->comparer = $comparer ?: function ($x, $y) {
-            return $x == $y;
-        };
-
+        $this->comparer = $comparer;
         $this->keySelector = $keySelector;
-
     }
 
     /**
@@ -43,15 +38,22 @@ class DistinctOperator implements OperatorInterface
                 try {
                     $key = $this->keySelector ? call_user_func($this->keySelector, $value) : $value;
 
-                    foreach ($values as $v) {
-                        $comparerEquals = call_user_func($this->comparer, $key, $v);
+                    if ($this->comparer) {
+                        foreach ($values as $v) {
+                            $comparerEquals = call_user_func($this->comparer, $key, $v);
 
-                        if ($comparerEquals) {
+                            if ($comparerEquals) {
+                                return;
+                            }
+                        }
+                        $values[] = $key;
+                    } else {
+                        if (array_key_exists($key, $values)) {
                             return;
                         }
+                        $values[$key] = null;
                     }
 
-                    $values[] = $key;
                     $observer->onNext($value);
 
                 } catch (\Exception $e) {