I've mentioned it in #134 already. This is a script runner with a couple of benchmarks that follow their RxJS 5 equivalents.

It can run all tests in the `/benchmark/**/*.php` directory with:
```
php benchmark/run.php
```

Or a single script:
```
php benchmark/run.php benchmark/distinct/distinct.php
```