@davidwdan 2017-02-09
@martinsik This looks awesome!  I ran it against v2 and I'm getting some unexpected results.  I would think that the performance would be about the same, but v1 is way faster for some of the operators!

v1:

```
Benchmarking 6 file/s (min 5s each)
script_name - total_runs (single_run_mean ±standard_variant)
============================================================
bufferWithCount - 107188 (0.05ms ±0.02ms)
distinct - 94617 (0.05ms ±0.03ms)
filter - 51071 (0.10ms ±0.03ms)
skipLast - 5888 (0.85ms ±0.16ms)
takeLast - 6514 (0.77ms ±0.14ms)
zip - 43674 (0.11ms ±0.03ms)
============================================================
total duration: 30.16s

```

v2:
```
Benchmarking 6 file/s (min 5s each)
script_name - total_runs (single_run_mean ±standard_variant)
============================================================
bufferWithCount - 42064 (0.12ms ±2.84ms)
distinct - 17970 (0.29ms ±6.12ms)
filter - 8470 (0.60ms ±10.11ms)
skipLast - 12612 (0.41ms ±9.08ms)
takeLast - 11036 (0.47ms ±10.37ms)
zip - 4856 (1.03ms ±16.54ms)
============================================================
total duration: 30.67s

```