@davidwdan 2017-02-09
Figured it out.  It's because we changed the default scheduler in v2 the EventLoopScheduler.  When I set them to Immediate, I get results closer to what I was expecting:

```
Benchmarking 6 file/s (min 5s each)
script_name - total_runs (single_run_mean ±standard_variant)
============================================================
bufferWithCount - 113451 (0.04ms ±0.02ms)
distinct - 116431 (0.04ms ±0.02ms)
filter - 59151 (0.08ms ±0.03ms)
skipLast - 6476 (0.77ms ±0.13ms)
takeLast - 7155 (0.70ms ±0.12ms)
zip - 47154 (0.11ms ±0.03ms)
============================================================
total duration: 30.16s

```

I take it that the plan is to eventually add benchmarks for the different schedulers.