@davidwdan 2017-02-09
With the latest changes:

v1:
```
Benchmarking 12 file/s (min 5s each)
script_name - total_runs (single_run_mean ±standard_deviation)
==============================================================
bufferWithCount - 107068 (0.05ms ±0.01ms)
bufferWithCount_eventloop - 6104 (0.82ms ±2.07ms)
distinct - 90765 (0.06ms ±0.05ms)
distinct_eventloop - 5935 (0.84ms ±2.05ms)
filter - 49457 (0.10ms ±0.03ms)
filter_eventloop - 2788 (1.80ms ±3.27ms)
skipLast - 5347 (0.94ms ±0.41ms)
skipLast_eventloop - 296 (16.89ms ±8.25ms)
takeLast - 6159 (0.81ms ±0.16ms)
takeLast_eventloop - 307 (16.32ms ±8.26ms)
zip - 42594 (0.12ms ±0.03ms)
zip_eventloop - 2846 (1.76ms ±2.42ms)
============================================================
total duration: 60.19s
```

v2:

```
Benchmarking 12 file/s (min 5s each)
script_name - total_runs (single_run_mean ±standard_deviation)
==============================================================
bufferWithCount - 118147 (0.04ms ±0.01ms)
bufferWithCount_eventloop - 30381 (0.17ms ±3.47ms)
distinct - 102737 (0.05ms ±0.39ms)
distinct_eventloop - 12915 (0.40ms ±7.58ms)
filter - 56806 (0.09ms ±0.14ms)
filter_eventloop - 6564 (0.77ms ±12.08ms)
skipLast - 6257 (0.80ms ±0.14ms)
skipLast_eventloop - 9151 (0.57ms ±11.23ms)
takeLast - 6927 (0.72ms ±0.13ms)
takeLast_eventloop - 7902 (0.65ms ±12.83ms)
zip - 47426 (0.11ms ±0.02ms)
zip_eventloop - 3609 (1.42ms ±20.08ms)
============================================================
total duration: 60.85s
```