@martinsik 2017-02-09
It didn't occur to me at first but I think we'll have to write scripts for both `ImmediateScheduler` and `EventLoopScheduler` similarly to https://github.com/ReactiveX/rxjs/tree/master/perf/micro in RxJS 5.

I updated `run.php` to also allow running all scripts in a single directory for practical reasons:
```
php benchmark/run.php benchmark/skipLast/
```