@martinsik 2017-03-18
I updated the benchmark runner script so when using a custom loop it doesn't staring the loop over and over again. The results should be more accurate:

Before:

```
Benchmarking 12 file/s (min 5s each)
script_name - total_runs (single_run_mean ±standard_deviation) - mem_start [mem_100_iter] mem_end
==============================================================
bufferWithCount - 33946 (0.15ms ±0.05ms) - 1.11MB 1.15MB 3.25MB 
bufferWithCount_eventloop - 2523 (1.98ms ±3.14ms) - 1.25MB 3.75MB 6.60MB 
distinct - 32050 (0.16ms ±0.12ms) - 6.47MB 6.53MB 3.33MB 
distinct_eventloop - 2627 (1.90ms ±2.94ms) - 2.28MB 6.21MB 5.22MB 
filter - 16977 (0.29ms ±0.08ms) - 5.09MB 5.11MB 6.15MB 
filter_eventloop - 1319 (3.79ms ±4.23ms) - 5.10MB 3.87MB 2.23MB 
skipLast - 1948 (2.57ms ±0.46ms) - 2.17MB 2.18MB 2.24MB 
skipLast_eventloop - 125 (40.04ms ±11.27ms) - 2.18MB 3.49MB 5.14MB 
takeLast - 2142 (2.33ms ±0.39ms) - 5.13MB 5.15MB 5.28MB 
takeLast_eventloop - 131 (38.20ms ±12.35ms) - 5.15MB 6.50MB 5.59MB 
zip - 14106 (0.35ms ±0.10ms) - 5.58MB 5.61MB 6.13MB 
zip_eventloop - 1704 (2.93ms ±2.59ms) - 5.61MB 2.07MB 5.21MB 
============================================================
total duration: 60.22s
```

After:
```
Benchmarking 12 file/s (min 5s each)
script_name - total_runs (single_run_mean ±standard_deviation) - mem_start [mem_100_iter] mem_end
==============================================================
bufferWithCount - 28074 (0.16ms ±0.30ms) - 1.14MB 1.16MB 2.20MB 
bufferWithCount_eventloop - 7383 (0.83ms ±1.56ms) - 1.25MB 4.16MB 4.18MB 6.30MB 
distinct - 25544 (0.17ms ±0.16ms) - 6.30MB 6.36MB 2.63MB 
distinct_eventloop - 26442 (0.16ms ±0.10ms) - 1.59MB 1.64MB 3.29MB 
filter - 16575 (0.27ms ±0.10ms) - 3.30MB 3.31MB 4.35MB 
filter_eventloop - 4037 (1.71ms ±2.22ms) - 3.31MB 5.55MB 
skipLast - 1814 (2.66ms ±0.34ms) - 5.55MB 5.57MB 5.63MB 
skipLast_eventloop - 419 (18.88ms ±26.40ms) - 5.56MB 2.89MB 
takeLast - 2215 (2.16ms ±0.29ms) - 2.90MB 2.91MB 3.04MB 
takeLast_eventloop - 443 (17.67ms ±24.99ms) - 2.91MB 6.11MB 
zip - 14025 (0.33ms ±0.08ms) - 6.12MB 6.14MB 6.66MB 
zip_eventloop - 3625 (1.36ms ±1.79ms) - 6.14MB 2.64MB 1.96MB 
============================================================
total duration: 60.26s
```